respond to more feedback from Gorry Fairhurst
Add a few statements describing the types of events that are logged.

Clarify the Security Considerations section.

>> ----------------------------------------------------------------------
>> DISCUSS:
>> ----------------------------------------------------------------------
 [...]
>> 2. If a packet is not matched and is dropped, ought it to be logged? (It would
>> be logged if there was no FID). Either way, the specification ought to explain
>> whether an entry in the log is to be expected.
> Currently, the cases where a log entry should be made are those where a
> flow control data item is received, by either side, that contains a FID
> not previously defined by the modem.  There is no logging based on what
> is seen in the dataplane.  Should this still be called out explicitly?

Understood. I do think it would help to just explain the logging
entries. Knowing what is being logged can be very helpful when comparing
implementations.

>> ----------------------------------------------------------------------
>> COMMENT:
>> ----------------------------------------------------------------------
>>
>> I don’t think the security mechanisms are sufficiently defined to be useful.
>> Please can you explain what threats these updates are expected to counter. This
>> currently leads to two very ineffective security considerations:
> The Security Considerations give an example of the type of threat to
> which flow control might susceptible.  Other cases are given in DLEP
> (8175) and are applicable, but not specific to flow control.
>
> Your last sentence ends with a colon, so I'm not sure if your comment
> was truncated.

The truncated part was only the text: "The transport layer security
mechanisms documented in [RFC8175] can, with some updates, be applied to
this document." (There is a similar phrase related to Ethernet below
this text also.)

I think the above current I-D text is likely just too terse for me to
understand if there is a missing something or not, especially the phrase
"with some updates", which made me wonder what sort of updates does this
imply ?

Whereas, the text you wrote above is much clearer to me, and I think
would help others if in the document.